Explain correctly why setup is trustless
Erase that Alice is protected against Bob not broadcasting F because
she holds C1a and RD1a, write that C1a and RD1a ensure her that her
funds will not be locked in the 2-of-2 multisig of F in case Bob does
not cooperate. Clarify that if Bob does not cooperate, nothing has
changed on the blockchain, so Alice is risking nothing.