Add that channel capacity must be verified
In case of Lightning multi-hop payments, the information on the
blockchain is not enough in order to be sure that the existing channels
can carry the required funds, but parties of these channels should
verify that they can complete the payment given the current state of
their channel. Thanks to Aydin Abadi for pointing this out.